refactor(example): split the fixture into a demo tier and lib/scenarios/
The example's lib/ had grown to 26 flat files, most of them regression
fixtures for one detection rule each, indistinguishable at a glance from the
seven files the README actually walks through.

The demo files stay at lib/; every rule-specific fixture moves under
lib/scenarios/. The default-run assertions now exclude them with one
`lib/scenarios/**` glob instead of a 15-entry list that had to be extended by
hand for every new fixture, and example/README.md documents both tiers.

Also drops two fixtures that duplicated coverage:

- `SilentSignal` in guards.dart, whose test asserted the same guard and the
  same outcome as `EmptyableStatus` — the two differed only in whether a return
  type or a parameter type kept the enum alive, which the guard does not read.
- `DeadLeafWidget` in widgets.dart: with no State pairing it is an ordinary
  dead class, already covered by `FullyDeadClass` in lib/orphans.dart.

The README's expected output had drifted from what the tool prints (it predates
the doc-only section, the dead-class reporting, and the operator default), so
it is refreshed from a real run, and `buildSerializable` no longer feeds
`jsonEncode` an object it cannot serialize, making the demo app runnable again.
