+++ example/lib/scenarios/comment_annotations.dart
