+++ example/lib/scenarios/enum_values.dart
