+++ example/lib/scenarios/freezed_unions.dart
