+++ example/lib/scenarios/freezed_unions.freezed.dart
