+++ example/lib/scenarios/serialization.freezed.dart
