+++ example/lib/scenarios/unions.dart
