+++ example/lib/scenarios/xref_analytics.dart
@@ -1,7 +1,7 @@
 // Provides the API the consumer uses, plus one normal reference that keeps
 // XrefEvent.signIn used.
 
-import 'package:sample_pkg/xref_event.dart';
+import 'package:sample_pkg/scenarios/xref_event.dart';
 
 class XrefAnalytics {
   void logEvent(XrefEvent event) => print(event);
