+++ example/lib/scenarios/xref_collision.dart
