+++ example/lib/scenarios/xref_collision_uses.dart
@@ -1,4 +1,4 @@
-import 'package:sample_pkg/xref_collision.dart';
+import 'package:sample_pkg/scenarios/xref_collision.dart';
 
 String describe(CollisionState state) => switch (state) {
   LiveState(status: true) => 'live',
