+++ example/lib/scenarios/xref_event.dart
