+++ example/lib/scenarios/xref_shapes.dart
