+++ example/lib/scenarios/xref_uses.dart
@@ -1,8 +1,8 @@
 // Deliberately does NOT import xref_event.dart; the reference below still
 // resolves without it.
 
-import 'package:sample_pkg/xref_analytics.dart';
-import 'package:sample_pkg/xref_shapes.dart';
+import 'package:sample_pkg/scenarios/xref_analytics.dart';
+import 'package:sample_pkg/scenarios/xref_shapes.dart';
 
 String describeShape(XrefState state) => switch (state) {
   XrefLoadedState(hasActive: true, ready: true) => 'both',
